[Narwhal] split LeaderSchedule into a separate file
